Rename Cell to CellFabric to improve readability
