+++ mathesar_ui/src/components/cell-fabric/CellFabric.svelte
@@ -1,9 +1,7 @@
 <!--
   @component
 
-  TODO: Move this component within Sheet
-
-  This component is meant to be common for tables, views, and for import preview
+  This component is meant to be common for tables, queries, and for import preview
 -->
 <script lang="ts">
   import type { ProcessedColumn } from '@mathesar/stores/table-data/processedColumns';
@@ -19,7 +17,7 @@
   $: props = cellComponentAndProps.props as Record<string, unknown>;
 </script>
 
-<div class="sheet-cell" data-column-id={processedColumn.column.id}>
+<div class="cell-fabric" data-column-id={processedColumn.column.id}>
   <svelte:component
     this={component}
     {...props}
@@ -37,7 +35,7 @@
 </div>
 
 <style lang="scss">
-  .sheet-cell {
+  .cell-fabric {
     --cell-height: 29px;
     position: relative;
     display: flex;
@@ -47,7 +45,7 @@
     width: 100%;
   }
 
-  .sheet-cell {
+  .cell-fabric {
     .loader {
       top: 6px;
       left: 8px;
