+++ mathesar_ui/src/components/cell-fabric/data-types/components/__tests__/SteppedInputCell.test.ts
