+++ mathesar_ui/src/components/cell-fabric/data-types/components/formatted-input/FormattedInputCell.svelte
