+++ mathesar_ui/src/components/cell-fabric/data-types/components/linked-record/LinkedRecordCell.svelte
