+++ mathesar_ui/src/components/cell-fabric/data-types/components/money/MoneyCellInput.svelte
