+++ mathesar_ui/src/components/cell-fabric/data-types/components/number/NumberCellInput.svelte
