+++ mathesar_ui/src/components/cell-fabric/data-types/components/uri/UriCell.svelte
