+++ mathesar_ui/src/components/cell-fabric/data-types/number.ts
