+++ mathesar_ui/src/stores/abstract-types/types.ts
@@ -4,7 +4,7 @@ import type {
   FormValues,
   IconProps,
 } from '@mathesar-component-library/types';
-import type { CellDataType } from '@mathesar/components/cell/data-types/typeDefinitions';
+import type { CellDataType } from '@mathesar/components/cell-fabric/data-types/typeDefinitions';
 import type { DbType } from '@mathesar/AppTypes';
 import type { Column } from '@mathesar/api/tables/columns';
 import type { States } from '@mathesar/utils/api';
